+++ lib/veteran_enrollment_system/enrollment_periods/service.rb
@@ -3,7 +3,6 @@
 require 'common/client/base'
 require 'common/client/concerns/monitoring'
 require 'veteran_enrollment_system/enrollment_periods/configuration'
-require 'veteran_enrollment_system/error_handling'
 
 module VeteranEnrollmentSystem
   module EnrollmentPeriods
@@ -18,10 +17,17 @@ module VeteranEnrollmentSystem
     #
     class Service < Common::Client::Base
       include Common::Client::Concerns::Monitoring
-      include VeteranEnrollmentSystem::ErrorHandling
 
       configuration VeteranEnrollmentSystem::EnrollmentPeriods::Configuration
       STATSD_KEY_PREFIX = 'api.enrollment_periods'
+      ERROR_MAP = {
+        400 => Common::Exceptions::BadRequest,
+        403 => Common::Exceptions::Forbidden,
+        404 => Common::Exceptions::ResourceNotFound,
+        500 => Common::Exceptions::ExternalServerInternalServerError,
+        502 => Common::Exceptions::BadGateway,
+        504 => Common::Exceptions::GatewayTimeout
+      }.freeze
 
       # Fetch enrollment periods by ICN from the enrollment system
       #
@@ -45,6 +51,25 @@ module VeteranEnrollmentSystem
           "get_enrollment_periods failed: #{e.respond_to?(:errors) ? e.errors.first[:detail] : e.message}"
         )
         raise e
+      end
+
+      private
+
+      # Raises mapped error, logs, and increments StatsD for error cases.
+      # Optionally accepts a statsd_key_prefix and operation name for metrics/logging.
+      def raise_error(response, statsd_key_prefix: nil, operation: nil)
+        message = response.body['messages']&.pluck('description')&.join(', ') || response.body
+
+        # if statsd_key_prefix && operation
+        #   # StatsD.increment("#{statsd_key_prefix}.#{operation}.failed")
+        #   Rails.logger.error(
+        #     "#{operation} failed: #{message}"
+        #   )
+        # end
+
+        # Just in case the status is not in the ERROR_MAP, raise a BackendServiceException
+        raise ERROR_MAP[response.status]&.new(detail: message) ||
+              Common::Exceptions::BackendServiceException.new(nil, detail: message)
       end
     end
   end
